Changed the Divide-method so that it returns double

+++ src/main/java/com/example/springdocker/MyMathCalculator.java
@@ -17,7 +17,7 @@ public class MyMathCalculator {
         return a * b;
     }
 
-    public float divide(float a, float b) {
+    public double divide(double a, double b) {
         return a / b;
     }
 }
